Moves pipeline ligoc into package
